Remove copy of old gtest.
The copy of gtest in the repository was old enough that it broke with
the new compiler on Ubuntu 14.04.  Rather than try to keep the copy of
gtest up to date, I deleted it and changed the build scripts to expect
it to be installed on the system at GTEST_DIR, a directory specified as
a build command line argument.

I also updated the README to reflect this change.