Modify build system to use user-specified gtest.
The copy of gtest in the repository was old enough that it broke with
the new compiler on Ubuntu 14.04. Rather than try to keep the copy of
gtest up to date, change the build scripts to expect it to be installed
on the system at GTEST_DIR, a directory specified as a build command
line argument.

Also update the README to reflect this change.